+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
  - Fix issues with PID Conditional Actions ([#416](https://github.com/kizniche/mycodo/issues/416))
  - Fix display of last edge on Live page
+ - Fix issue updating the status of some dependencies after their install
 
 ### Miscellaneous